Do not run document-start scripts every time the tab is focused.
Using .addProgressListener() will fire every time the value in the address bar changes.  Good if you want to act on that, but bat for this use case.  On the other hand, .addTabsProgressListener() filters down to an individual tab's activity (not the entire browser), so doesn't consider tab focus, and doesn't fire then when we don't want it.

Also clean up the noisy un-used parameters to all the other callback functions.