Use coral cdn when checking userscripts.org for updates.
Side effect: reverted the "update URL must be secure" commit.  We need to not have that for this to work, and in thinking about this I realized that isn't necessary.  (Even if you can fake the update check, the actual update's download still needs to be secure.)
Also: Refactored updateURL so the patching logic is in the getter, thus always applied regardless of order of operations.