Convert http-USO to https-USO to enable updating
Several scripts are stuck at an old version because their downloadURL
points to the http-version of USO instead of https, while
extension.greasemonkey.requireSecureUpdates is set to true by default.

Since it is known in advance that USO is able to server content over
https, we can safely convert all USO downloadURLs to the https-version.