Create sandboxes with sameZoneAs to reduce GC costs
by default sandboxes are created in the system zone which has a lifetime
equal to that of the process and has a fairly large footprint.

Since GM sandboxes are scoped to individual content pages they should be
allocated in the same GC zone as the content. This allows them to be
discarded together and also reduces the footprint (and thus GC overhead)
for the system zone.