Show script source: Switch to `.addTab()`.
Something deep in Firefox internals was failing, given the way we
called `.loadOneTab()`.  Switch to `.addTab()` and tweak the
notification bar to (apparently?) cope with the different timing
that results.

Along the way: fix the is-temp-script method to actually be handled,
by listening for it via the PPMM, not global MM.

Fixes #2202