Allow more contentType when detecting scripts
application/javascript is the "correct" mime for it. But some others may also be accepted by browser. HTML specification had given a list of mime which may be considered as javascript file at https://html.spec.whatwg.org/multipage/scripting.html#javascript-mime-type . some of them just trigger a download action on firefox. we do not include theme. text/plain should not be readed as javascript, but it may be useful for install scripts from github with raw button.