package.sh: use bash directly for getting version
package.sh tries to extract the version from manifest.json.
Unfortunately the method used is specific to GNU sed, so the command
fails on OS X and other operating systems with the BSD userland.

Since we're already depending on bash, just use bash directly to extract
the version.  It isn't pretty but avoids external dependencies (besides
zip, of course).