Move strings into locales.
Use a pattern like gettext() where the original string stays in the source, so the source is still readable.
This means a lot of copy/paste into _locales/en/messages.json.

Include a few strings in German, because I can speak it _just_ enough to test in German.  Just to be sure it's working.

Fixes #2527