Add accessibility semantics and labels.
This makes Greasemonkey far more usable by screen reader users.

Even though keyboard navigation is implemented for the Monkey Menu, this change exposes it as buttons, rather than menu + menuitems.
This is because unlike most menus, this menu contains headings, which doesn't fit the ARIA menu pattern.
This way, screen reader users will be able to browse this like any other document and will be made aware of (and able to jump to) the headings.