Add an optional simple editor which uses a plain textarea.
This is needed by screen reader users, as CodeMirror is unfortunately not accessible.
Having a separate editor is not ideal, nor is the missing functionality for screen reader users.
Nevertheless, this is far better than a completely inaccessible editor.
This at least allows screen reader users to read the content and do basic editing.