Allow editing (text) resources.
We already saved the mimetype.  In the previous commit, we added URL -- because URL is a required input to the downloader, used to "install" the newly edited script (and, download things if the script was edited to reference a new require/resource/icon).
Now for each resource, if mimetype is text add a tab to edit it, and process this tab when saving.

Fixes #2733