Do not process messages before scripts are loaded.
This resolves a long standing (but minor until recently) race condition.  (Almost) all of the messages that the background script processes depend on already having loaded the installed user scripts.  If the message is sent before this happens (because it's async.) it will not be serviced correctly.  This was rare in the past, but after 69de9382 this will be common, as the background script is non-persistent.

So, A) add a promise object, B) manually resolve it after the whole script loading process is complete, C) in the "on message" handler, await completion of that process before calling the message handler.

Refs #3173