Fix issue preventing some scripts from auto-updating.
GreaseMonkey computes the next scheduled update for each script each time it successfully automatically checks for an update. This information is used to select the next script which automatically check for an update. This change ensures that scripts which do not have this information stored (i.e., scripts which have yet to have a successful automatic update check) will perform their initial automatic update check regardless of the order in which scripts are iterated.

Due to the existing parameters and tuning in GreaseMonkey, update-eligible scripts which have not had their updates checked will cause GreaseMonkey to pause updates for 3 hours prior to making the first update check. You must ensure that you keep your browser opene for at least 3 hours (plus 15 minutes) for each script which has not had an automatic update check performed on it before other scripts can resume their normal update schedule.

See #3190 and the discussion of updateNextAt.